polling service: move some classes to src/service
